Update to run SPA null model only once per GWAS
Runs epactsSingleNull.R to create the null object for SPA test before starting the GWAS, so that the null model is not needed to be calculated for every chunk (of size "unit", default 500K bp) of SNPs. After running epactsSingleNull.R, each instance of epactsSingle.R just loads the null object and performs the test. This helps to reduce the chunk size (memory efficient) and increase the number of chunks without requiring to calculate the null object for each chunk.